所有的表数据都能上传

+++ src/models/_entities/system_admin.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "ty_system_admin")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i16,
     pub account: String,
     pub pwd: String,
     pub real_name: String,
